quest enemies helper method
