+++ src/core/quest/index.ts
@@ -3,3 +3,4 @@ export * from './getRequiredQuests'
 export * from './getRequiredSkills'
 export * from './getRecommendedItems'
 export * from './getRecommendedSkills'
+export * from './getEnemiesToKill'
