+++ src/tools/wiki/getQuestInfo.ts
@@ -9,7 +9,7 @@ import { extractTemplatesFromXML as extractTemplatesFromXML } from '../../utils/
 import { findTemplates } from '../../utils/templateHelpers.js';
 import { QuestInfoToolResponse } from '../../zod';
 import { QuestInfoToolResponseType } from '../../types/osrsMcp.js';
-import { getRequiredItems, getRequiredQuests, getRequiredSkills, getRecommendedItems, getRecommendedSkills } from '../../core/quest';
+import { getRequiredItems, getRequiredQuests, getRequiredSkills, getRecommendedItems, getRecommendedSkills, getEnemiesToKill } from '../../core/quest';
 
 export async function getQuestInfo(
 	questName: string,
@@ -54,46 +54,9 @@ export async function getQuestInfo(
 		response.recommendedSkills = getRecommendedSkills(recommended)
 	}
 
-	// Parse enemies to defeat from the kills string
-	const enemiesToDefeat: Record<string, { levels: number[] }> = {}
 	if (kills) {
-		// Split by lines to process each enemy entry
-		const lines = kills.split('\n')
-
-		for (const line of lines) {
-			// Match patterns like "* [[Slagilith]] ''(level 92)''" or "* [[Dwarf gang member]]s ''(level 44/48/49)'' ([[Multicombat area]])"
-			// The pattern looks for lines starting with asterisks followed by [[Enemy Name]]
-			const enemyMatch = line.match(/^\*+\s*\[\[([^\]]+)\]\]/)
-			if (enemyMatch) {
-				const enemyNameRaw = enemyMatch[1]
-				// If there's a pipe, the actual enemy name is before it
-				const enemyName = enemyNameRaw.split('|')[0].trim()
-
-				if (enemyName) {
-					// Extract level information from the line
-					// Look for patterns like ''(level 92)'' or ''(level 44/48/49)''
-					const levelMatch = line.match(/''?\(level\s+([^)]+)\)''?/)
-					const levels: number[] = []
-
-					if (levelMatch) {
-						const levelStr = levelMatch[1]
-						// Handle multiple levels separated by slashes
-						const levelParts = levelStr.split('/')
-						for (const part of levelParts) {
-							const level = parseInt(part.trim(), 10)
-							if (!isNaN(level)) {
-								levels.push(level)
-							}
-						}
-					}
-
-					// If we found levels, use them; otherwise default to empty array
-					enemiesToDefeat[enemyName] = { levels }
-				}
-			}
-		}
+		response.enemiesToDefeat = getEnemiesToKill(kills)
 	}
-	response.enemiesToDefeat = enemiesToDefeat
 
 	const { name, number, image, release, update, aka, members, series, developer } = infoboxQuestTemplate.parameters
